requirements.txt
asgiref==3.8.1
attrs==23.2.0
config==0.5.1
diff-match-patch==20230430
Django==5.0.6
django-extensions==3.2.3
django-import-export==4.0.3
djangorestframework==3.15.1
drf-yasg==1.21.7
import-export==0.3.1
inflection==0.5.1
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==69.5.1
sqlparse==0.5.0
tablib==3.5.0
types-pkg-resources==0.1.3
uritemplate==4.1.1
